internal: add ci exports
