+++ .blueprint/from-issue/generator.ts
@@ -3,8 +3,7 @@ import { join } from 'node:path';
 import EnvironmentBuilder from '../../cli/environment-builder.ts';
 import BaseGenerator from '../../generators/base-core/index.ts';
 import { type InfoData, extractDataFromInfo, markdownDetails } from '../../generators/info/support/index.ts';
-import { appendToSummary, getGithubIssue, setGithubTaskOutput } from '../../lib/testing/github.ts';
-import { prepareSample } from '../../lib/testing/sample-config.ts';
+import { appendToSummary, getGithubIssue, prepareSample, setGithubTaskOutput } from '../../lib/ci/index.ts';
 import { promptSamplesFolder } from '../support.ts';
 
 const YO_RC_OUTPUT = 'yo-rc';
