+++ .github/workflows/issue-check.yml
@@ -71,7 +71,7 @@ jobs:
       #----------------------------------------------------------------------
       # Install JHipster and generate project+entities
       #----------------------------------------------------------------------
-      - run: npm ci --ignore-scripts
+      - run: npm ci --prod --ignore-scripts
         working-directory: ${{ github.workspace }}/generator-jhipster
       - uses: jhipster/actions/build-jhipster-bom@v1
         with:
