+++ .github/workflows/vue.yml
@@ -56,7 +56,7 @@ jobs:
       - uses: actions/setup-node@v6
         with:
           node-version: ${{ steps.node-version.outputs.node-version }}
-      - run: npm ci --ignore-scripts
+      - run: npm ci --prod --ignore-scripts
       - id: build
         run: bin/jhipster.cjs github-build-matrix vue --event-name ${{ github.event_name }}
   applications:
@@ -102,7 +102,7 @@ jobs:
       #----------------------------------------------------------------------
       # Install JHipster and generate project+entities
       #----------------------------------------------------------------------
-      - run: npm ci --ignore-scripts
+      - run: npm ci --prod --ignore-scripts
         working-directory: ${{ github.workspace }}/generator-jhipster
       - name: 'GENERATION: project'
         run: jhipster.cjs generate-sample ${{ matrix.sample }} --skip-jhipster-dependencies --skip-checks --skip-install --no-insight
