+++ lib/ci/apply-patch-to-template.ts
