+++ lib/testing/index.ts
@@ -1,16 +1,10 @@
-export { default as getGenerator } from './get-generator.ts';
 export * from './helpers.ts';
-export * from './github.ts';
-export * from './github-group.ts';
-export * from './github-matrix.ts';
-export * from './sample-config.ts';
-
 export * from './mutate-data.ts';
 
 // test matrix
-export * from './support/application-samples.ts';
-export * from './support/client-samples.ts';
-export * from './support/entity-samples.ts';
-export * from './support/matrix-utils.ts';
 export * from './support/matcher.ts';
-export * from './support/server-samples.ts';
+export * from './support/entity-samples.ts';
+export * from '../ci/support/application-samples.ts';
+export * from '../ci/support/client-samples.ts';
+export * from '../ci/support/matrix-utils.ts';
+export * from '../ci/support/server-samples.ts';
