+++ test/support/tests.ts
@@ -10,7 +10,8 @@ import type CoreGenerator from '../../generators/base-core/index.ts';
 import { CONTEXT_DATA_APPLICATION_KEY } from '../../generators/base-simple-application/support/constants.ts';
 import { WORKSPACES_PRIORITY_NAMES } from '../../generators/base-workspaces/priorities.ts';
 import { GENERATOR_JHIPSTER } from '../../generators/generator-constants.ts';
-import { getGenerator, runResult, skipPrettierHelpers as helpers } from '../../lib/testing/index.ts';
+import { runResult, skipPrettierHelpers as helpers } from '../../lib/testing/index.ts';
+import getGenerator from '../../lib/utils/get-generator.ts';
 
 const workspacesPriorityList = Object.values(WORKSPACES_PRIORITY_NAMES);
